Using floating point type in programme.

+++ main.c
@@ -6,16 +6,18 @@
 
 int main()
 {
-    int fahr, celsius;
+    float fahr, celsius;
     int lower, upper, step;
     lower = 0; /* lower boundary for temperature table */
     upper = 300; /* upper boundary */
     step = 20; /* step */
     fahr = lower;
-    while (fahr <= upper) {
-        celsius = 5 * (fahr-32) / 9;
-        printf ("%d\t%d\n", fahr, celsius);
+    while (fahr <= upper)
+    {
+        celsius = 5 * (fahr - 32) / 9;
+        printf("%f\t%f\n", fahr, celsius);
         fahr = fahr + step;
     }
+
     return 0;
 }
